getUrlContent should verify the SSL certificate
In case of TLS connections is `getUrlContent` currently not verifying the SSL certificate if `curl` is not available.

Additionally as a second layer of defense `CURLOPT_SSL_VERIFYPEER = true` is enforced, this should be anyways the default since cURL 7.10.

Make it possible to disable certificate check

There may be some reason where a SSL certificate check is not required, e.g. for the "Get from URL" function.

However, this is subject to discussion and the certificate check is enabled by default.

Add verify cert function

Also update the public interface

make false default for  getUrlContent's verifyCert

add folder for certs

ssl improvements